+++ lib/src/database/units/han_solo.dart
@@ -7,7 +7,7 @@ final hanSolo = Unit(
   id: 'han-solo',
   name: 'Han Solo',
   subTitle: 'Unorthodox General',
-  faction: Faction.lightSide,
+  faction: Faction.rebels,
   isUnique: true,
   miniatures: 1,
   rank: Rank.commander,