+++ lib/src/database/upgrades/force/force_choke.dart
@@ -5,7 +5,7 @@ final forceChoke = Upgrade(
   name: 'Force Choke',
   type: UpgradeSlot.force,
   isExhaustible: true,
-  restrictedToFaction: Faction.darkSide,
+  restrictedToForceAlignment: ForceAlignment.dark,
   points: 5,
   text: r'''
     Choose an enemy trooper mini at range 1 that is not a commander.